python/requirements.txt
flask==0.12.0
django==1.11.0
pyyaml==3.12
requests==2.6.0
jinja2==2.7.2
werkzeug==0.11.0
sqlalchemy==1.0.0
pillow==2.0.0
cryptography==1.0
paramiko==1.15.0
lxml==3.4.0
beautifulsoup4==4.6.0
urllib3==1.24.1
pycrypto==2.6.1
pickle5==0.0.8
